+++ common/src/main/java/com/like/common/view/bottomnavigationbar/BottomTabInfoOrigin.java
@@ -70,7 +70,7 @@ public class BottomTabInfoOrigin extends BottomTabInfo {
         mBadgeViewHelper.setTextColor(color);
     }
 
-    public void setMessageTextSize(float size) {
+    public void setMessageTextSize(int size) {
         mBadgeViewHelper.setTextSize(size);
     }
